pyproject.toml
[project]
name = "pleiades"
version = "0.1.0"
description = ""
authors = [
    {name = "Angelo Alvisi",email = "angeloalvisi@gmail.com"}
]
readme = "README.md"
requires-python = "<4.0.0,>=3.10"
dependencies = [
    "chainlit (>=2.5.5,<3.0.0)",
    "google-genai (>=1.20.0,<2.0.0)",
    "google-api-python-client (>=2.172.0,<3.0.0)",
    "ollama (>=0.5.1,<0.6.0)"
]

[tool.poetry]
packages = [{include = "pleiades", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
